[CLI]: Loading modules from diff paths
This commit introduces a feature to include a module from a special
directory mentioned in the config.ini. (Config.ini has a default
variable for path_to_suite)

On the basis of that the the river_core.py was modified to include
modules from abs path. And some minor exit message enhacements.

To manage the path of the module, a variable similar to outputdir was
included in the sim_hookspecs.py.